Deflate pkg/policy into the pkg/pluginsdk/policy package
